qcacmn: Correct the data type of byte count variable
Use a 64-bit data type for byte count variable to prevent
data overflow.

CRs-Fixed: 2165595

+++ dp/inc/cdp_txrx_cmn_struct.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011-2017 The Linux Foundation. All rights reserved.
+ * Copyright (c) 2011-2018 The Linux Foundation. All rights reserved.
  *
  * Permission to use, copy, modify, and/or distribute this software for
  * any purpose with or without fee is hereby granted, provided that the
@@ -787,7 +787,7 @@ struct cdp_pkt_info {
 	/*no of packets*/
 	uint32_t num;
 	/* total no of bytes */
-	uint32_t bytes;
+	uint64_t bytes;
 };
 
 /* Tx  Stats */
